# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserProfile": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"phone"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProfileRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UserSettingsModelType": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"settingsValue": {
						"type": "string"
					},
					"settingsKey": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"settingsValue",
					"settingsKey"
				],
				"type": "object",
				"description": "User Settings"
			},
...
			"UpdateSettingRequest": {
				"properties": {
					"value": {
						"type": "string"
					}
				},
				"required": [
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "core-api",
		"version": "1.4.0",
		"description": "API documentation for the KiranaPro - 'PROD' Backend microservice, designed for seamless integration with multiple microservices, including AI, ONDC/ KIKO/ third-party ONDC services, and mobile frontends (iOS & Android).",
		"license": {
			"name": "Proprietary - Copyright (c) 2024, KiranaPro"
		},
		"contact": {
			"name": "Sandeep K ",
			"email": "sandeep@kirana.pro"
		}
	},
	"paths": {
		"/user": {
			"post": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "The updated profile.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedProfile": {
											"$ref": "#/components/schemas/UserProfile"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"updatedProfile",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Invalid inputs"
					}
				},
				"description": "Update the user's profile settings.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The updated profile data (name and/or email).",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProfileRequest",
								"description": "The updated profile data (name and/or email)."
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "The user's profile.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserProfile"
								}
							}
						}
					}
				},
				"description": "Get the user's profile settings.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/user/settings": {
			"get": {
				"operationId": "GetUserSettings",
				"responses": {
					"200": {
						"description": "An array of user settings.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserSettingsModelType"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieve all user settings.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUserSetting",
				"responses": {
					"200": {
						"description": "A success message.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Invalid inputs"
					}
				},
				"description": "Create a new user setting.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The key and value of the setting.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSettingRequest",
								"description": "The key and value of the setting."
							}
						}
					}
				}
			}
		},
		"/user/settings/{key}": {
			"put": {
				"operationId": "UpdateUserSetting",
				"responses": {
					"200": {
						"description": "A success message or an error if the setting is not found.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Setting not found"
					}
				},
				"description": "Update an existing user setting.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The key of the setting to update.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The updated value.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSettingRequest",
								"description": "The updated value."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUserSetting",
				"responses": {
					"200": {
						"description": "A success message or an error if the setting is not found.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Setting not found"
					}
				},
				"description": "Delete a user setting.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The key of the setting to delete.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/address": {
			"get": {
				"operationId": "GetUserAddresses",
				"responses": {
					"200": {
						"description": "An array of user addresses.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserAddressModelType"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all user addresses.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUserAddress",
				"responses": {
					"200": {
						"description": "A success message and the created address.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreatedAddressResponse"
								}
							}
						}
					},
					"201": {
						"description": "Missing or invalid inputs"
					},
					"400": {
						"description": "Missing or invalid inputs"
					}
				},
				"description": "Create a new user address.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The address details.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserAddressRequest",
								"description": "The address details."
							}
						}
					}
				}
			}
		},
		"/user/address/{id}": {
			"get": {
				"operationId": "GetUserAddressById",
				"responses": {
					"200": {
						"description": "The address or an error if not found.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"address": {
											"$ref": "#/components/schemas/UserAddressModelType"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Address not found"
					}
				},
				"description": "Get a specific user address by ID.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The address ID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUserAddress",
				"responses": {
					"200": {
						"description": "A success message or an error if the address is not found.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"address": {
											"$ref": "#/components/schemas/UserAddressModelType"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Address not found"
					}
				},
				"description": "Delete a user address by ID.",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The address ID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/tasks": {
			"post": {
				"operationId": "CreateTask",
				"responses": {
					"200": {
						"description": "Task ID of the created task",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"task_id": {
											"type": "string"
										}
									},
									"required": [
										"task_id"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input"
					}
				},
				"description": "Create a new task for a user.",
				"tags": [
					"Task"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Task creation data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTaskRequest",
								"description": "Task creation data"
							}
						}
					}
				}
			}
		},
		"/tasks/{taskId}": {
			"get": {
				"operationId": "GetTask",
				"responses": {
					"200": {
						"description": "Task status or details",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskResponse"
								}
							}
						}
					},
					"404": {
						"description": "Task not found"
					}
				},
				"description": "Fetch the status or details of a task based on task ID.",
				"tags": [
					"Task"
				],
				"security": [],
				"parameters": [
					{
						"description": "Task ID to retrieve",
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/payments": {
			"post": {
				"operationId": "CreatePayment",
				"responses": {
					"200": {
						"description": "The Razorpay order details.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RazorPayOrderResponse"
								}
							}
						}
					},
					"400": {
						"description": "Missing or invalid inputs"
					}
				},
				"description": "Create a new Razorpay payment order.",
				"tags": [
					"Payments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"amount": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"amount"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/payments/verify": {
			"post": {
				"operationId": "VerifyPayment",
				"responses": {
					"200": {
						"description": "A success message or an error.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"order": {},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Missing or invalid inputs"
					}
				},
				"description": "Verify a Razorpay payment and create an order.",
				"tags": [
					"Payments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The request body with payment and order details.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyPaymentBody",
								"description": "The request body with payment and order details."
							}
						}
					}
				}
			}
		},
		"/orders": {
			"get": {
				"operationId": "GetOrders",
				"responses": {
					"200": {
						"description": "A list of orders.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"orders": {
											"items": {
												"$ref": "#/components/schemas/OrderResponse"
											},
											"type": "array"
										}
									},
									"required": [
										"orders"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Fetch all orders for the logged-in user.",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"200": {
						"description": "The created order.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"order": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"order",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "CartId and AddressId are required"
					}
				},
				"description": "Create a new order.",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The request body containing cartId and addressId.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"addressId": {
										"type": "string"
									},
									"cartId": {
										"type": "string"
									}
								},
								"required": [
									"addressId",
									"cartId"
								],
								"type": "object",
								"description": "The request body containing cartId and addressId."
							}
						}
					}
				}
			}
		},
		"/orders/{id}": {
			"get": {
				"operationId": "GetOrder",
				"responses": {
					"200": {
						"description": "The order details.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrderResponse"
								}
							}
						}
					}
				},
				"description": "Get a specific order by ID.",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the order to fetch.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orders/{id}/track": {
			"get": {
				"operationId": "TrackOrder",
				"responses": {
					"200": {
						"description": "The order's tracking status.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"deliveryStatus": {
											"type": "string"
										},
										"orderStatus": {
											"type": "string"
										}
									},
									"required": [
										"deliveryStatus",
										"orderStatus"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get the tracking status of a specific order.",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the order to track.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orders/{id}/cancel": {
			"post": {
				"operationId": "CancelOrder",
				"responses": {
					"200": {
						"description": "A success message or an error.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Cancel an order.",
				"tags": [
					"Orders"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the order to cancel.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/notifications": {
			"get": {
				"operationId": "GetNotifications",
				"responses": {
					"200": {
						"description": "A list of notifications.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"notifications": {
											"items": {
												"$ref": "#/components/schemas/Notification"
											},
											"type": "array"
										}
									},
									"required": [
										"notifications"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "No notifications found"
					}
				},
				"description": "Fetch all notifications for the logged-in user.",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateNotification",
				"responses": {
					"200": {
						"description": "A success message with the created notification.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateNotificationResponse"
								}
							}
						}
					},
					"201": {
						"description": "Notification added successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateNotificationResponse"
								}
							}
						}
					},
					"400": {
						"description": "Missing or invalid inputs"
					}
				},
				"description": "Create a new notification.",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The request body containing the notification details.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateNotificationRequest",
								"description": "The request body containing the notification details."
							}
						}
					}
				}
			}
		},
		"/notifications/{id}": {
			"delete": {
				"operationId": "DeleteNotification",
				"responses": {
					"200": {
						"description": "A success message with the deleted notification details.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"notification": {
											"$ref": "#/components/schemas/Notification"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"notification",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Notification not found"
					}
				},
				"description": "Delete a notification by ID.",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the notification to delete.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/home": {
			"get": {
				"operationId": "GetHome",
				"responses": {
					"200": {
						"description": "Home page details",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"carousels": {
											"properties": {
												"title": {
													"type": "string"
												},
												"data": {
													"items": {
														"properties": {
															"deeplink": {
																"type": "string"
															},
															"image_url": {
																"type": "string"
															},
															"id": {
																"type": "string"
															}
														},
														"required": [
															"deeplink",
															"image_url",
															"id"
														],
														"type": "object"
													},
													"type": "array"
												}
											},
											"required": [
												"title",
												"data"
											],
											"type": "object"
										},
										"banners": {
											"properties": {
												"title": {
													"type": "string"
												},
												"data": {
													"items": {
														"properties": {
															"deeplink": {
																"type": "string"
															},
															"image_url": {
																"type": "string"
															},
															"id": {
																"type": "string"
															}
														},
														"required": [
															"deeplink",
															"image_url",
															"id"
														],
														"type": "object"
													},
													"type": "array"
												}
											},
											"required": [
												"title",
												"data"
											],
											"type": "object"
										},
										"welcomeSubText": {
											"type": "string"
										},
										"welcomeText": {
											"type": "string"
										}
									},
									"required": [
										"carousels",
										"banners",
										"welcomeSubText",
										"welcomeText"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Retrieves home page data, including welcome text, banners, and carousels.",
				"tags": [
					"Home"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
		"/audio": {
			"post": {
				"operationId": "ProcessAudio",
				"responses": {
					"200": {
						"description": "A task ID for the processed audio",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AudioProcessResponse"
								}
							}
						}
					}
				},
				"description": "Process an uploaded audio file and return a task ID.",
				"tags": [
					"Audio"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The User-Agent header for device identification",
						"in": "header",
						"name": "user-agent",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary",
										"description": "The uploaded MP3 audio file"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/v1"
		}
	]
}